cors cookie 처리. 로그인 성공

+++ api/controllers/utils.go
@@ -8,6 +8,6 @@ import (
 func userIdFromContext(ctx echo.Context) int64 {
 	t := ctx.Get("user").(*jwt.Token)
 	claims := t.Claims.(jwt.MapClaims)
-	uid := claims["uid"].(int64)
-	return uid
+	uid := claims["uid"].(float64) // int64로 값을 넣어줬음에도 float64 로 돌려준다. json 파싱 라이브러리 때문인듯
+	return int64(uid)
 }
